OLCand rest endpoint: move all classes in de.metas.rest_api.ordercandidates.* packages
* also change class access level from public to default/package level.
* dropped ProductMasterDataContext - not used

https://github.com/metasfresh/metasfresh/issues/5469
