+++ de.metas.business.rest-api-impl/src/main/java/de/metas/rest_api/ordercandidates/impl/BPartnerMasterDataProvider.java
@@ -74,7 +74,7 @@ import lombok.Value;
  * #L%
  */
 
-public class BPartnerMasterDataProvider
+final class BPartnerMasterDataProvider
 {
 	public static BPartnerMasterDataProvider of(
 			@Nullable final Properties ctx,
@@ -116,7 +116,7 @@ public class BPartnerMasterDataProvider
 		return handleBPartnerInfoWithContext(jsonBPartnerInfo, context);
 	}
 
-	public final BPartnerInfo getCreateBPartnerInfo(
+	public BPartnerInfo getCreateBPartnerInfo(
 			@Nullable final JsonBPartnerInfo jsonBPartnerInfo,
 			final OrgId orgId)
 	{
@@ -190,7 +190,7 @@ public class BPartnerMasterDataProvider
 		}
 	}
 
-	private final BPartnerInfo lookupBPartnerInfoOrNull(
+	private BPartnerInfo lookupBPartnerInfoOrNull(
 			@NonNull final JsonBPartnerInfo jsonBPartnerInfo,
 			@NonNull final BPartnerMasterDataContext context)
 	{
@@ -217,7 +217,7 @@ public class BPartnerMasterDataProvider
 				.build();
 	}
 
-	private final BPartnerId lookupBPartnerIdOrNull(
+	private BPartnerId lookupBPartnerIdOrNull(
 			@NonNull final JsonBPartnerInfo jsonBPartnerInfo,
 			@NonNull final BPartnerMasterDataContext context)
 	{
@@ -251,7 +251,7 @@ public class BPartnerMasterDataProvider
 				.orElse(null);
 	}
 
-	private final BPartnerLocationId lookupBPartnerLocationIdOrNull(
+	private BPartnerLocationId lookupBPartnerLocationIdOrNull(
 			@NonNull final JsonRequestLocation jsonBPartnerLocation,
 			@NonNull final BPartnerMasterDataContext context)
 	{
@@ -363,7 +363,7 @@ public class BPartnerMasterDataProvider
 		return BPartnerId.ofRepoId(bpartnerRecord.getC_BPartner_ID());
 	}
 
-	private final void updateBPartnerRecord(
+	private void updateBPartnerRecord(
 			@NonNull final I_C_BPartner bpartnerRecord,
 			@NonNull final JsonRequestBPartner from)
 	{
