+++ de.metas.business.rest-api-impl/src/main/java/de/metas/rest_api/ordercandidates/impl/ProductNotFoundException.java
@@ -1,9 +1,6 @@
-package de.metas.rest_api.product.impl;
+package de.metas.rest_api.ordercandidates.impl;
 
-import de.metas.organization.OrgId;
-import lombok.Builder;
-import lombok.NonNull;
-import lombok.Value;
+import org.adempiere.exceptions.AdempiereException;
 
 /*
  * #%L
@@ -15,27 +12,23 @@ import lombok.Value;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
-@Value
-@Builder(toBuilder = true)
-public class ProductMasterDataContext
+@SuppressWarnings("serial")
+class ProductNotFoundException extends AdempiereException
 {
-	public static ProductMasterDataContext ofOrg(@NonNull final OrgId orgId)
+	public ProductNotFoundException(String msg)
 	{
-		return builder().orgId(orgId).build();
+		super(msg);
 	}
-
-	@NonNull
-	OrgId orgId;
 }
