+++ de.metas.business.rest-api/src/main/java/de/metas/rest_api/ordercandidates/JsonOLCandCreateRequest.java
@@ -12,7 +12,6 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import de.metas.rest_api.SyncAdvise;
-import de.metas.rest_api.product.JsonProductInfo;
 import de.metas.util.Check;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Builder;
