+++ vertical-healthcare_ch/forum_datenaustausch_ch.invoice_rest-api/src/main/java/de/metas/vertical/healthcare/forum_datenaustausch_ch/rest/XmlToOLCandsService.java
@@ -46,10 +46,10 @@ import de.metas.rest_api.ordercandidates.JsonOLCandCreateBulkRequest;
 import de.metas.rest_api.ordercandidates.JsonOLCandCreateBulkResponse;
 import de.metas.rest_api.ordercandidates.JsonOLCandCreateRequest;
 import de.metas.rest_api.ordercandidates.JsonOLCandCreateRequest.JsonOLCandCreateRequestBuilder;
+import de.metas.rest_api.ordercandidates.JsonProductInfo.Type;
 import de.metas.rest_api.ordercandidates.JsonOrganization;
+import de.metas.rest_api.ordercandidates.JsonProductInfo;
 import de.metas.rest_api.ordercandidates.OrderCandidatesRestEndpoint;
-import de.metas.rest_api.product.JsonProductInfo;
-import de.metas.rest_api.product.JsonProductInfo.Type;
 import de.metas.util.Check;
 import de.metas.util.StringUtils;
 import de.metas.util.collections.CollectionUtils;
